(chore) Clean up lint errors
Moved one_of and issue to m.core since those are made from the fp and issue modules. This created circular dependencies.
