+++ packages/python/m/core/json.py
@@ -2,8 +2,9 @@ import json
 import sys
 from typing import Any, Optional, Mapping as Map, List, cast, Union
 from collections.abc import Mapping
+from . import issue
 from .fp import Good, OneOf
-from .issue import Issue, issue
+from .issue import Issue
 from .io import CITool
