+++ packages/python/m/github/api.py
@@ -1,6 +1,7 @@
 from typing import Mapping, Any
-from ..core.fp import OneOf, Good, one_of
-from ..core.issue import Issue, issue
+from ..core import one_of, issue
+from ..core.fp import OneOf, Good
+from ..core.issue import Issue
 from ..core.http import fetch_json
