+++ packages/python/tests/ci/config.py
@@ -2,7 +2,8 @@ import unittest
 from unittest.mock import patch
 from typing import cast
 
-from m.core.issue import Issue, issue
+from m.core import issue
+from m.core.issue import Issue
 from m.core.fp import Good
 from m.ci.config import read_config, Config
